refactor(sections): extract the bare-word block scanners into scan.rs
Pure motion: operand tables, argparse subparser lists, headingless
invocation tables and bare command tables, with their tests.
